Fixed Damage application bug
Forgot to inherit from Idamage interface.
added a reticle to the UI.
Changed takeDamage parameter to a float
changed Coins to a float
changed addcoins parameter to float

+++ Assets/Scripts/GameManager.cs
@@ -16,7 +16,7 @@ public class GameManager : MonoBehaviour
     public GameObject looseMenu;
 
     public int jumpCost;
-    public int coins;
+    public float coins;
     public bool isPaused;
     float timeScaleOrig;
     public GameObject playerSpawnPos;
@@ -50,7 +50,7 @@ public class GameManager : MonoBehaviour
         }
     }
 
-    public void addCoins(int amount)
+    public void addCoins(float amount)
     {
         coins += amount;
     }
